Add files via upload
"I modified the colors in my code to ensure that it is accessible to people with color blindness. I used a color palette called 'glasbey_dark' from the 'colorcet' library, which has been specifically designed for colorblindness. In addition, I made sure to maintain the order of colors in the original code to ensure that the meaning of the data remains the same"